% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/morphism.R
\name{adjoint}
\alias{adjoint}
\alias{adjoint,Morphism-method}
\alias{adjoint,IdentityMorphism-method}
\alias{adjoint,Affine3DMorphism-method}
\alias{adjoint,Warp3DMorphism-method}
\alias{adjoint,VolToSurfMorphism-method}
\alias{adjoint,SurfToSurfMorphism-method}
\title{Get adjoint (generalized inverse)}
\usage{
adjoint(object, ...)

\S4method{adjoint}{Morphism}(object, ...)

\S4method{adjoint}{IdentityMorphism}(object, ...)

\S4method{adjoint}{Affine3DMorphism}(object, ...)

\S4method{adjoint}{Warp3DMorphism}(object, ...)

\S4method{adjoint}{VolToSurfMorphism}(object, ...)

\S4method{adjoint}{SurfToSurfMorphism}(object, ...)
}
\arguments{
\item{object}{A Morphism object}

\item{...}{Additional arguments}
}
\value{
Adjoint morphism or error if not available
}
\description{
Returns an adjoint morphism when defined.
}
\details{
For invertible morphisms, \code{adjoint()} is defined and equals \code{invert()}.
For other morphisms, \code{adjoint()} may not be implemented and will error.
}
