% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{Affine3DMorphism-class}
\alias{Affine3DMorphism-class}
\title{3D affine morphism}
\description{
Represents an affine transformation (rotation, translation, scaling, shear)
between volume spaces.
}
\section{Slots}{

\describe{
\item{\code{matrix}}{4x4 affine transformation matrix (maps target -> source)}
}}

