% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_utils.R
\name{affine_from_components}
\alias{affine_from_components}
\title{Wrap an affine matrix built from components into an Affine3DMorphism}
\usage{
affine_from_components(source, target, ...)
}
\description{
Wrap an affine matrix built from components into an Affine3DMorphism
}
