% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_warp.R
\name{afni_is_oblique}
\alias{afni_is_oblique}
\title{Detect whether an affine is oblique (not cardinal axis-aligned)}
\usage{
afni_is_oblique(affine, threshold_deg = 0.01)
}
\arguments{
\item{affine}{4x4 voxel-to-world affine}

\item{threshold_deg}{Angular threshold in degrees}
}
\value{
Logical
}
\description{
Detect whether an affine is oblique (not cardinal axis-aligned)
}
