% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warp_chain.R
\name{apply_warp_chain}
\alias{apply_warp_chain}
\title{Apply a warp chain to coordinates}
\usage{
apply_warp_chain(chain, coords)
}
\arguments{
\item{chain}{Chain object from build_warp_chain()}

\item{coords}{Numeric matrix (N x 3) of target world coords}
}
\value{
Transformed coordinates in source space
}
\description{
Applies an optimized warp chain built by build_warp_chain().
}
\keyword{internal}
