% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_utils.R
\name{check_slots}
\alias{check_slots}
\title{Check if object has required slots}
\usage{
check_slots(object, slots)
}
\arguments{
\item{object}{S4 object to check}

\item{slots}{Character vector of required slot names}
}
\value{
TRUE if valid, character error message otherwise
}
\description{
Check if object has required slots
}
\keyword{internal}
