% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_utils.R
\name{compute_hash}
\alias{compute_hash}
\title{Compute a stable hash for objects}
\usage{
compute_hash(..., algo = "xxhash64")
}
\arguments{
\item{...}{Objects to hash}

\item{algo}{Hash algorithm (default "xxhash64" for speed)}
}
\value{
Character string hash
}
\description{
Creates a reproducible hash using digest with serialize=TRUE.
}
\examples{
compute_hash("hello", "world")
compute_hash(list(a = 1, b = 2))
}
