% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_absolute_to_displacement}
\alias{cpp_absolute_to_displacement}
\title{Convert absolute-coordinate warp field to displacement field}
\usage{
cpp_absolute_to_displacement(field_abs, dim, vox_to_world)
}
\arguments{
\item{field_abs}{Absolute coordinate field (flattened)}

\item{dim}{Field dimensions}

\item{vox_to_world}{Voxel-to-world affine}
}
\value{
Displacement field
}
\description{
Converts a warp field storing absolute world coordinates to displacement field.
}
\keyword{internal}
