% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_apply_affine_chain}
\alias{cpp_apply_affine_chain}
\title{Apply chain of affine transformations to coordinates}
\usage{
cpp_apply_affine_chain(coords, matrices)
}
\arguments{
\item{coords}{Numeric matrix (N x 3) of coordinates}

\item{matrices}{List of 4x4 affine matrices}
}
\value{
Transformed coordinates
}
\description{
Efficiently applies multiple 4x4 affine matrices to coordinates.
}
\keyword{internal}
