% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_apply_projector}
\alias{cpp_apply_projector}
\title{Apply sparse projector matrix to data}
\usage{
cpp_apply_projector(projMat, data, threads = 1L)
}
\arguments{
\item{projMat}{dgCMatrix sparse projector}

\item{data}{Numeric matrix of data}

\item{threads}{Number of OpenMP threads}
}
\value{
Projected data matrix
}
\description{
Fast C++ sparse matrix multiplication.
}
\keyword{internal}
