% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_apply_warp_field}
\alias{cpp_apply_warp_field}
\title{Apply warp displacement field to coordinates}
\usage{
cpp_apply_warp_field(coords, field, dim, voxel_to_world)
}
\arguments{
\item{coords}{Numeric matrix (N x 3) of world coordinates}

\item{field}{Numeric vector of displacement values (flattened 3 x X x Y x Z)}

\item{dim}{Integer vector of dimensions (X, Y, Z)}

\item{voxel_to_world}{4x4 matrix (actually world_to_voxel inverse)}
}
\value{
Numeric matrix (N x 3) of transformed coordinates
}
\description{
Fast C++ implementation of trilinear interpolation for displacement fields.
}
\keyword{internal}
