% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_path_apply_affine_warp_affine}
\alias{cpp_path_apply_affine_warp_affine}
\title{Apply affines -> warp -> affines to coordinates (target -> source)}
\usage{
cpp_path_apply_affine_warp_affine(
  coords,
  affines_pre,
  field,
  dim,
  world_to_vox,
  cubic,
  affines_post
)
}
\arguments{
\item{coords}{n x 3 target world coords}

\item{affines_pre}{list of 4x4 matrices applied before warp}

\item{field}{displacement field}

\item{dim}{field dims}

\item{world_to_vox}{warp world_to_vox}

\item{cubic}{use cubic sampling}

\item{affines_post}{list of 4x4 matrices applied after warp}
}
\value{
n x 3 source world coords
}
\description{
Apply affines -> warp -> affines to coordinates (target -> source)
}
