% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_path_apply_steps}
\alias{cpp_path_apply_steps}
\title{Apply arbitrary steps (affine/warp) to coordinates (target -> source)}
\usage{
cpp_path_apply_steps(coords, steps)
}
\arguments{
\item{coords}{n x 3 target world coords}

\item{steps}{list of steps with kind \"affine\" or \"warp\" and associated data}
}
\value{
n x 3 transformed coords
}
\description{
Apply arbitrary steps (affine/warp) to coordinates (target -> source)
}
