% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_ribbon_sample_volume}
\alias{cpp_ribbon_sample_volume}
\title{Ribbon sampling from volume onto surface}
\usage{
cpp_ribbon_sample_volume(data, inner, outer, world_to_vox, n_steps, method)
}
\arguments{
\item{data}{Volume data (3D array)}

\item{inner}{Inner surface coordinates (N x 3)}

\item{outer}{Outer surface coordinates (N x 3)}

\item{world_to_vox}{4x4 world-to-voxel transform}

\item{n_steps}{Number of samples along ribbon}

\item{method}{Interpolation method}
}
\value{
Numeric vector of N averaged values
}
\description{
Samples along cortical ribbon (inner to outer surface) and averages.
}
\keyword{internal}
