% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_triplets_to_dgC}
\alias{cpp_triplets_to_dgC}
\title{Fast triplet to dgCMatrix assembly with duplicate aggregation}
\usage{
cpp_triplets_to_dgC(i, j, x, nrow, ncol, threads = 1L)
}
\arguments{
\item{i}{1-based row indices}

\item{j}{1-based col indices}

\item{x}{values}

\item{nrow}{number of rows}

\item{ncol}{number of cols}

\item{threads}{number of threads (unused currently)}
}
\value{
dgCMatrix sparse matrix
}
\description{
Converts triplet format (i, j, x) to sparse column matrix.
}
\keyword{internal}
