% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_volume_world_coords}
\alias{cpp_volume_world_coords}
\title{Generate world coordinates for full volume grid}
\usage{
cpp_volume_world_coords(dim, affine)
}
\arguments{
\item{dim}{integer vector length 3}

\item{affine}{4x4 voxel-to-world}
}
\value{
Numeric matrix (N x 3) of world coordinates
}
\description{
Generate world coordinates for full volume grid
}
\keyword{internal}
