% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{extract_affine}
\alias{extract_affine}
\alias{extract_affine,array-method}
\alias{extract_affine,matrix-method}
\alias{extract_affine,ANY-method}
\alias{extract_affine,DenseNeuroVol-method}
\alias{extract_affine,DenseNeuroVec-method}
\title{Extract affine matrix from data}
\usage{
extract_affine(x)

\S4method{extract_affine}{array}(x)

\S4method{extract_affine}{matrix}(x)

\S4method{extract_affine}{ANY}(x)

\S4method{extract_affine}{DenseNeuroVol}(x)

\S4method{extract_affine}{DenseNeuroVec}(x)
}
\arguments{
\item{x}{Data object}
}
\value{
4x4 affine matrix
}
\description{
Extract affine matrix from data
}
