% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_ingest.R
\name{fsl_fsl_to_world}
\alias{fsl_fsl_to_world}
\title{Convert FSL coords to world coords}
\usage{
fsl_fsl_to_world(affine)
}
\arguments{
\item{affine}{4x4 voxel-to-world affine}
}
\value{
4x4 FSL-to-world transform
}
\description{
Convert FSL coords to world coords
}
\keyword{internal}
