% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_ingest.R
\name{fsl_load_flirt_morphism}
\alias{fsl_load_flirt_morphism}
\title{Build Affine3DMorphism from FLIRT matrix}
\usage{
fsl_load_flirt_morphism(
  source,
  target,
  mat_path,
  cost = 1,
  method_tag = "anatomical"
)
}
\arguments{
\item{source}{Source domain (with @geometry@affine) or affine matrix}

\item{target}{Target domain (with @geometry@affine) or affine matrix}

\item{mat_path}{Path to FLIRT .mat file}

\item{cost}{Path cost}

\item{method_tag}{Method tag}
}
\value{
Affine3DMorphism object
}
\description{
Creates a morphism from an FSL FLIRT/MCFLIRT matrix file.
}
\keyword{internal}
