% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_ingest.R
\name{fsl_spacing_from_affine}
\alias{fsl_spacing_from_affine}
\title{Compute voxel spacing from affine}
\usage{
fsl_spacing_from_affine(affine)
}
\arguments{
\item{affine}{4x4 voxel-to-world affine}
}
\value{
Numeric vector of spacing in each dimension
}
\description{
Compute voxel spacing from affine
}
\keyword{internal}
