% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_ingest.R
\name{fsl_world_to_fsl}
\alias{fsl_world_to_fsl}
\title{Convert world coords to FSL coords}
\usage{
fsl_world_to_fsl(affine, dim = NULL)
}
\arguments{
\item{affine}{4x4 voxel-to-world affine}

\item{dim}{Optional image dimensions}
}
\value{
4x4 world-to-FSL transform
}
\description{
Convert world coords to FSL coords
}
\keyword{internal}
