% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warp_loader.R
\name{get_loader}
\alias{get_loader}
\alias{get_warp_loader}
\title{Retrieve a warp loader by name}
\usage{
get_loader(name = NULL)

get_warp_loader(name = NULL)
}
\arguments{
\item{name}{Loader name (default "rnifti")}
}
\value{
Loader function
}
\description{
Gets a loader function from the registry.
}
\examples{
loader <- get_loader("rnifti")
}
