% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\docType{class}
\name{Grid-class}
\alias{Grid-class}
\alias{show,Grid-method}
\title{Grid: regular coordinate grid specification}
\usage{
\S4method{show}{Grid}(object)
}
\arguments{
\item{object}{A Grid object (for show method)}
}
\description{
Grid: regular coordinate grid specification
}
\section{Slots}{

\describe{
\item{\code{dims}}{Integer: grid dimensions}

\item{\code{affine}}{4x4 voxel-to-world matrix}
}}

