% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{grid_coords}
\alias{grid_coords}
\title{Get world coordinates for a grid}
\usage{
grid_coords(grid)
}
\arguments{
\item{grid}{Grid object}
}
\value{
Numeric matrix (prod(dims) x 3)
}
\description{
Get world coordinates for a grid
}
