% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{grid_from_data}
\alias{grid_from_data}
\title{Get grid specification from volume data}
\usage{
grid_from_data(data)
}
\arguments{
\item{data}{Volume with geometry}
}
\value{
Grid object
}
\description{
Get grid specification from volume data
}
