% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_helpers.R
\name{grid_of}
\alias{grid_of}
\title{Extract Grid from volume or grid-like input}
\usage{
grid_of(x)
}
\arguments{
\item{x}{neuroim2 volume, array, or Grid}
}
\value{
Grid
}
\description{
Extract Grid from volume or grid-like input
}
