% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{IdentityMorphism-class}
\alias{IdentityMorphism-class}
\title{Identity morphism}
\description{
The identity morphism maps a domain to itself. Required for categorical
structure and useful for composition.
}
