% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/morphism.R
\name{invert}
\alias{invert}
\alias{invert,Morphism-method}
\alias{invert,IdentityMorphism-method}
\alias{invert,Affine3DMorphism-method}
\alias{invert,Warp3DMorphism-method}
\title{Invert a morphism}
\usage{
invert(object)

\S4method{invert}{Morphism}(object)

\S4method{invert}{IdentityMorphism}(object)

\S4method{invert}{Affine3DMorphism}(object)

\S4method{invert}{Warp3DMorphism}(object)
}
\arguments{
\item{object}{A Morphism object}
}
\value{
Inverted morphism
}
\description{
Returns the inverse morphism if available (inverse_type == "exact" or "approximate").
For morphisms with approximate or adjoint inverses, use adjoint() instead.
}
\examples{
aff <- Affine3DMorphism("a", "b", diag(4))
inv <- invert(aff)  # b -> a
}
