% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{invert_affine}
\alias{invert_affine}
\title{Invert an affine transformation}
\usage{
invert_affine(affine)
}
\arguments{
\item{affine}{4x4 affine matrix}
}
\value{
Inverted 4x4 affine matrix
}
\description{
Invert an affine transformation
}
\examples{
affine <- diag(4)
affine[1:3, 4] <- c(5, 10, 15)
inv <- invert_affine(affine)
}
