% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_utils.R
\name{is_affine_morphism}
\alias{is_affine_morphism}
\title{Check if an object is an Affine3DMorphism}
\usage{
is_affine_morphism(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical
}
\description{
Check if an object is an Affine3DMorphism
}
