% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphism.R
\name{is_linear_morphism}
\alias{is_linear_morphism}
\title{Check if morphism is linear (affine or identity)}
\usage{
is_linear_morphism(object)
}
\arguments{
\item{object}{A Morphism object}
}
\value{
Logical
}
\description{
Check if morphism is linear (affine or identity)
}
