% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphism.R
\name{is_valid_path}
\alias{is_valid_path}
\title{Check if a morphism path is valid}
\usage{
is_valid_path(path)
}
\arguments{
\item{path}{List of Morphism objects}
}
\value{
Logical
}
\description{
Returns TRUE if the path is composable, FALSE otherwise.
}
\examples{
aff1 <- Affine3DMorphism("a", "b", diag(4))
aff2 <- Affine3DMorphism("b", "c", diag(4))
is_valid_path(list(aff1, aff2))  # TRUE
}
