% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphism.R
\name{is_warp_morphism}
\alias{is_warp_morphism}
\title{Check if morphism is a warp field}
\usage{
is_warp_morphism(object)
}
\arguments{
\item{object}{A Morphism object}
}
\value{
Logical
}
\description{
Check if morphism is a warp field
}
