% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobian.R
\name{jacobian_det}
\alias{jacobian_det}
\alias{jacobian_det,IdentityMorphism-method}
\alias{jacobian_det,Affine3DMorphism-method}
\alias{jacobian_det,Warp3DMorphism-method}
\alias{jacobian_det,MorphismPath-method}
\title{Compute Jacobian determinants}
\usage{
jacobian_det(
  morphism,
  coords,
  log = FALSE,
  mode = c("pullback", "pushforward")
)

\S4method{jacobian_det}{IdentityMorphism}(
  morphism,
  coords,
  log = FALSE,
  mode = c("pullback", "pushforward")
)

\S4method{jacobian_det}{Affine3DMorphism}(
  morphism,
  coords,
  log = FALSE,
  mode = c("pullback", "pushforward")
)

\S4method{jacobian_det}{Warp3DMorphism}(
  morphism,
  coords,
  log = FALSE,
  mode = c("pullback", "pushforward")
)

\S4method{jacobian_det}{MorphismPath}(
  morphism,
  coords,
  log = FALSE,
  mode = c("pullback", "pushforward")
)
}
\arguments{
\item{morphism}{A Morphism or MorphismPath}

\item{coords}{Numeric matrix (n x 3) of coordinates}

\item{log}{Return log(|det(J)|)?}

\item{mode}{"pullback" or "pushforward"}
}
\value{
Numeric vector of determinants
}
\description{
Compute Jacobian determinants
}
\examples{
aff <- Affine3DMorphism("a", "b", diag(4) * 2)
coords <- matrix(c(0, 0, 0), ncol = 3)
jacobian_det(aff, coords)  # 8 (2^3)
jacobian_det(aff, coords, log = TRUE)  # log(8)
}
