% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deformation_field.R
\name{jacobian_det_field}
\alias{jacobian_det_field}
\title{Extract Jacobian determinant volume from a morphism on a grid}
\usage{
jacobian_det_field(morphism, grid, log = FALSE)
}
\arguments{
\item{morphism}{A Morphism (affine or warp)}

\item{grid}{A Grid object}

\item{log}{Logical; if TRUE, return log-Jacobian determinants}
}
\value{
Array with grid dimensions containing Jacobian determinants
}
\description{
Extract Jacobian determinant volume from a morphism on a grid
}
