% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warp_loader.R
\name{list_loaders}
\alias{list_loaders}
\title{List all registered loaders}
\usage{
list_loaders()
}
\value{
Character vector of registered loader names
}
\description{
List all registered loaders
}
\examples{
list_loaders()
}
