% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warp_loader.R
\name{load_warp_neuroim2}
\alias{load_warp_neuroim2}
\title{Default neuroim2 warp loader}
\usage{
load_warp_neuroim2(path)
}
\arguments{
\item{path}{Path to NIfTI warp file}
}
\value{
List with array, dim, world_to_vox, vox_to_world
}
\description{
Loads a NIfTI displacement field using neuroim2 (DenseNeuroVec).
}
\keyword{internal}
