% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{lps_to_ras}
\alias{lps_to_ras}
\title{Convert LPS coordinates to RAS}
\usage{
lps_to_ras(coords)
}
\arguments{
\item{coords}{Numeric matrix (N x 3) or vector of length 3 in LPS}
}
\value{
Coordinates in RAS convention
}
\description{
ANTs and ITK use LPS (Left-Posterior-Superior) convention. This function
converts to our internal RAS convention by flipping the first two axes.
}
\examples{
lps <- c(-10, -20, 30)
ras <- lps_to_ras(lps)
}
