% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{mesh_is_sphere}
\alias{mesh_is_sphere}
\title{Check whether a surface mesh is approximately spherical}
\usage{
mesh_is_sphere(mesh, tolerance = 1.001)
}
\arguments{
\item{mesh}{SurfaceMesh object}

\item{tolerance}{Ratio tolerance for min/max radius comparison}
}
\value{
Logical
}
\description{
Check whether a surface mesh is approximately spherical
}
