% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{mesh_set_radius}
\alias{mesh_set_radius}
\title{Set mesh radius for spherical meshes}
\usage{
mesh_set_radius(mesh, radius = 100)
}
\arguments{
\item{mesh}{SurfaceMesh object}

\item{radius}{Target radius}
}
\value{
SurfaceMesh object with rescaled coordinates
}
\description{
Set mesh radius for spherical meshes
}
