% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphism.R
\name{morphism_kind}
\alias{morphism_kind}
\title{Get the kind of a morphism}
\usage{
morphism_kind(object)
}
\arguments{
\item{object}{A Morphism object}
}
\value{
Character string indicating kind
}
\description{
Returns the morphism kind (identity, affine3d, warp3d, vol2surf, surf2surf).
}
\examples{
aff <- Affine3DMorphism("a", "b", diag(4))
morphism_kind(aff)  # "affine3d"
}
