% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_utils.R
\name{new_cache_env}
\alias{new_cache_env}
\title{Create a new environment for caching}
\usage{
new_cache_env()
}
\value{
A new environment with hash=TRUE for efficient lookup
}
\description{
Create a new environment for caching
}
\keyword{internal}
