% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_utils.R
\name{null-coalesce}
\alias{null-coalesce}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Left operand}

\item{y}{Right operand (used if x is NULL)}
}
\value{
x if not NULL, otherwise y
}
\description{
Null-coalescing operator
}
\keyword{internal}
