% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{ras_to_lps}
\alias{ras_to_lps}
\title{Convert RAS coordinates to LPS}
\usage{
ras_to_lps(coords)
}
\arguments{
\item{coords}{Numeric matrix (N x 3) or vector of length 3 in RAS}
}
\value{
Coordinates in LPS convention
}
\description{
Converts from our internal RAS convention to ANTs/ITK LPS convention.
}
\examples{
ras <- c(10, 20, 30)
lps <- ras_to_lps(ras)
}
