% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_helpers.R
\name{read_image}
\alias{read_image}
\title{Simple image I/O using neuroim2}
\usage{
read_image(path)
}
\arguments{
\item{path}{File path}
}
\value{
neuroim2 volume (DenseNeuroVol or DenseNeuroVec) or array fallback
}
\description{
Simple image I/O using neuroim2
}
