% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_io.R
\name{read_linear_transform_array}
\alias{read_linear_transform_array}
\title{Read a series of linear transforms}
\usage{
read_linear_transform_array(
  path,
  format = c("generic", "fsl", "afni", "itk", "lta", "x5"),
  source = "source",
  target = "target",
  ...
)
}
\arguments{
\item{path}{Base file path}

\item{format}{One of \code{"generic"}, \code{"fsl"}, \code{"afni"}, or \code{"itk"}}

\item{source}{Source domain id}

\item{target}{Target domain id}

\item{...}{Passed to \code{read_linear_transform()}}
}
\value{
A \code{LinearTransformArray} object
}
\description{
Supports single-file transforms and FSL/MCFLIRT-style indexed files
(\code{path.000}, \code{path.001}, ...).
}
