% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{resample_volume}
\alias{resample_volume}
\title{Resample volume to new geometry}
\usage{
resample_volume(data, morphism, target, method = "linear", modulate = "none")
}
\arguments{
\item{data}{Source volume}

\item{morphism}{Transform}

\item{target}{Target geometry (Grid or volume)}

\item{method}{Interpolation method}

\item{modulate}{Jacobian modulation}
}
\value{
Array with target dimensions
}
\description{
Convenience wrapper for common volume-to-volume resampling.
}
