% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sample_volume_on_surface}
\alias{sample_volume_on_surface}
\title{Sample volume on surface}
\usage{
sample_volume_on_surface(
  data,
  morphism,
  surface_coords = NULL,
  method = "linear"
)
}
\arguments{
\item{data}{Source volume}

\item{morphism}{VolToSurfMorphism}

\item{surface_coords}{Surface vertex coordinates (optional)}

\item{method}{Interpolation method}
}
\value{
Values at surface vertices
}
\description{
Project volumetric data onto surface vertices.
}
