% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\docType{class}
\name{SampledPoints-class}
\alias{SampledPoints-class}
\alias{show,SampledPoints-method}
\title{Sampled points (irregular spatial reference)}
\usage{
\S4method{show}{SampledPoints}(object)
}
\arguments{
\item{object}{A \code{SampledPoints} object (for the show method).}
}
\description{
Lightweight container for point samples in world coordinates.
}
\section{Slots}{

\describe{
\item{\code{coords}}{Numeric matrix (N x 3)}
}}

