% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\docType{class}
\name{Sampler-class}
\alias{Sampler-class}
\alias{show,Sampler-method}
\title{Sampler: data + interpolation encapsulation}
\usage{
\S4method{show}{Sampler}(object)
}
\arguments{
\item{object}{A Sampler object (for show method)}
}
\description{
Sampler: data + interpolation encapsulation
}
\section{Slots}{

\describe{
\item{\code{evaluate}}{Function: (coords) -> values}

\item{\code{dims}}{Integer: data dimensions}

\item{\code{vdim}}{Integer: output dimensionality per point}

\item{\code{bounds}}{List: min/max coordinate bounds}

\item{\code{method}}{Character: interpolation method}

\item{\code{outside}}{Value for out-of-bounds queries}
}}

