% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{surface_mesh}
\alias{surface_mesh}
\title{Construct a surface mesh}
\usage{
surface_mesh(vertices, faces = NULL, domain = NULL)
}
\arguments{
\item{vertices}{Numeric matrix with 3 columns (V x 3)}

\item{faces}{Optional matrix with 3 columns (F x 3), 0- or 1-based indices}

\item{domain}{Optional domain identifier}
}
\value{
SurfaceMesh object
}
\description{
Construct a surface mesh
}
