% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{tkras_to_ras}
\alias{tkras_to_ras}
\title{Convert FreeSurfer tkRAS to scanner RAS}
\usage{
tkras_to_ras(tkras, c_ras)
}
\arguments{
\item{tkras}{Numeric matrix (N x 3) or vector of length 3 in tkRAS}

\item{c_ras}{Numeric vector of length 3: the c_ras offset from FreeSurfer}
}
\value{
Coordinates in scanner RAS
}
\description{
FreeSurfer's tkregister uses a shifted coordinate system (tkRAS) that differs
from scanner RAS by the c_ras offset stored in the surface header.
}
\examples{
tkras <- c(0, 0, 0)
c_ras <- c(0.5, -2.3, 1.1)
scanner_ras <- tkras_to_ras(tkras, c_ras)
}
