% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/morphism.R
\name{transform}
\alias{transform}
\alias{transform,IdentityMorphism-method}
\alias{transform,Affine3DMorphism-method}
\alias{transform,Warp3DMorphism-method}
\alias{transform,VolToSurfMorphism-method}
\alias{transform,SurfToSurfMorphism-method}
\alias{transform,MorphismPath-method}
\title{Transform coordinates through a morphism (pullback)}
\usage{
transform(morphism, coords)

\S4method{transform}{IdentityMorphism}(morphism, coords)

\S4method{transform}{Affine3DMorphism}(morphism, coords)

\S4method{transform}{Warp3DMorphism}(morphism, coords)

\S4method{transform}{VolToSurfMorphism}(morphism, coords)

\S4method{transform}{SurfToSurfMorphism}(morphism, coords)

\S4method{transform}{MorphismPath}(morphism, coords)
}
\arguments{
\item{morphism}{A Morphism object or MorphismPath}

\item{coords}{Matrix of coordinates (N x 3) in target domain}
}
\value{
Matrix of coordinates (N x 3) in source domain
}
\description{
Applies the coordinate pullback: given coordinates in the target domain,
returns corresponding coordinates in the source domain.
}
\examples{
# Create an affine morphism
aff <- Affine3DMorphism("source_hash", "target_hash", diag(4))
coords <- matrix(c(10, 20, 30), ncol = 3)
transform(aff, coords)
}
