% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/morphism.R
\name{transform_coords}
\alias{transform_coords}
\alias{transform_coords,Morphism-method}
\alias{transform_coords,MorphismPath-method}
\title{Transform coordinates (compatibility alias)}
\usage{
transform_coords(object, coords)

\S4method{transform_coords}{Morphism}(object, coords)

\S4method{transform_coords}{MorphismPath}(object, coords)
}
\arguments{
\item{object}{A Morphism object}

\item{coords}{Matrix of coordinates (N x 3)}
}
\value{
Transformed coordinates
}
\description{
Alias for transform() to maintain neurofunctor compatibility.
}
