% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_io.R
\name{TransformFileError}
\alias{TransformFileError}
\title{Construct a transform file-format error condition}
\usage{
TransformFileError(message)
}
\arguments{
\item{message}{Error message}
}
\value{
Condition object of class \code{TransformFileError}
}
\description{
Construct a transform file-format error condition
}
