% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_io.R
\name{TransformIOError}
\alias{TransformIOError}
\title{Construct a transform IO error condition}
\usage{
TransformIOError(message)
}
\arguments{
\item{message}{Error message}
}
\value{
Condition object of class \code{TransformIOError}
}
\description{
Construct a transform IO error condition
}
