% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{VolToSurfMorphism-class}
\alias{VolToSurfMorphism-class}
\title{Volume-to-surface morphism}
\description{
Maps volume coordinates to surface vertices. Supports different sampling
strategies (ribbon, trilinear point sampling).
}
\section{Slots}{

\describe{
\item{\code{method}}{Sampling method ("ribbon", "trilinear", "nearest", "mid_thickness")}

\item{\code{ribbon_inner}}{Path to inner (white) surface for ribbon sampling}

\item{\code{ribbon_outer}}{Path to outer (pial) surface for ribbon sampling}

\item{\code{n_ribbon_samples}}{Number of samples along ribbon normal}
}}

