% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{voxels_to_world}
\alias{voxels_to_world}
\title{Voxel indices to world coordinates}
\usage{
voxels_to_world(voxels, affine)
}
\arguments{
\item{voxels}{Integer matrix (N x 3) of voxel indices (0-based)}

\item{affine}{4x4 voxel-to-world affine matrix}
}
\value{
Matrix of world coordinates (N x 3)
}
\description{
Convert voxel indices to world (RAS mm) coordinates using an affine.
}
\keyword{internal}
