% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warp_chain.R
\name{warp_chain_key}
\alias{warp_chain_key}
\title{Build cache key for a morphism path}
\usage{
warp_chain_key(path)
}
\arguments{
\item{path}{List of morphisms}
}
\value{
Character hash key
}
\description{
Build cache key for a morphism path
}
\keyword{internal}
