% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{world_to_voxels}
\alias{world_to_voxels}
\title{World coordinates to voxel indices}
\usage{
world_to_voxels(world, affine)
}
\arguments{
\item{world}{Numeric matrix (N x 3) of world coordinates}

\item{affine}{4x4 voxel-to-world affine matrix}
}
\value{
Matrix of voxel indices (0-based, continuous - round for discrete)
}
\description{
Convert world (RAS mm) coordinates to voxel indices using an affine.
}
\keyword{internal}
