% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_utils.R
\name{write_affine_matrix_txt}
\alias{write_affine_matrix_txt}
\title{Write a 4x4 affine matrix to a text file}
\usage{
write_affine_matrix_txt(matrix, file, type = NULL, comment = TRUE)
}
\description{
Write a 4x4 affine matrix to a text file
}
