% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_helpers.R
\name{write_image}
\alias{write_image}
\title{Write image using neuroim2}
\usage{
write_image(x, path)
}
\arguments{
\item{x}{neuroim2 volume or array}

\item{path}{Output path}
}
\description{
Write image using neuroim2
}
