% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_io.R
\name{write_linear_transform_array}
\alias{write_linear_transform_array}
\title{Write a series of linear transforms}
\usage{
write_linear_transform_array(
  x,
  path,
  format = c("generic", "fsl", "afni", "itk", "lta", "x5"),
  ...
)
}
\arguments{
\item{x}{A \code{LinearTransformArray}, list of affine morphisms/matrices, or single transform}

\item{path}{Base output path}

\item{format}{One of \code{"generic"}, \code{"fsl"}, \code{"afni"}, or \code{"itk"}}

\item{...}{Passed to \code{write_linear_transform()}}
}
\value{
Invisibly returns written path(s)
}
\description{
Write a series of linear transforms
}
