% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x5_io.R
\name{write_x5}
\alias{write_x5}
\title{Write X5 transform file}
\usage{
write_x5(path, x5_list)
}
\arguments{
\item{path}{Output file path}

\item{x5_list}{List of \code{X5Transform} objects}
}
\value{
Invisibly returns \code{path}
}
\description{
Write X5 transform file
}
