% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x5_io.R
\name{x5_domain}
\alias{x5_domain}
\title{Create an X5 domain descriptor}
\usage{
x5_domain(grid, size, mapping = NULL, coordinates = NULL)
}
\arguments{
\item{grid}{Logical}

\item{size}{Integer vector}

\item{mapping}{Optional matrix}

\item{coordinates}{Optional coordinate kind string}
}
\value{
An \code{X5Domain} object
}
\description{
Create an X5 domain descriptor
}
