% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_warp.R
\name{afni_aff12_to_ras}
\alias{afni_aff12_to_ras}
\title{Convert RAI affine to RAS}
\usage{
afni_aff12_to_ras(mat_rai)
}
\arguments{
\item{mat_rai}{4x4 affine in RAI coordinates}
}
\value{
4x4 affine in RAS coordinates
}
\description{
Converts an AFNI RAI affine matrix to our internal RAS convention
by flipping the Z axis.
}
\examples{
mat_rai <- diag(4)
mat_ras <- afni_aff12_to_ras(mat_rai)
}
