% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_warp.R
\name{afni_read_aff12}
\alias{afni_read_aff12}
\title{Read AFNI .aff12.1D affine matrix}
\usage{
afni_read_aff12(path)
}
\arguments{
\item{path}{Path to .aff12.1D file}
}
\value{
4x4 affine matrix (in RAI)
}
\description{
Reads a 3x4 affine matrix in RAI coordinates and returns a 4x4 matrix.
}
\examples{
\dontrun{
mat_rai <- afni_read_aff12("transform.aff12.1D")
}
}
