% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_helpers.R
\name{as_morphism_path}
\alias{as_morphism_path}
\title{Coerce list/paths to MorphismPath}
\usage{
as_morphism_path(...)
}
\arguments{
\item{...}{Morphism objects or file paths to transforms}
}
\value{
MorphismPath object
}
\description{
Coerce list/paths to MorphismPath
}
