% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_barycentric_sample}
\alias{cpp_barycentric_sample}
\title{Barycentric interpolation on surface mesh}
\usage{
cpp_barycentric_sample(coords, vertices, faces, data)
}
\arguments{
\item{coords}{Query coordinates (N x 3)}

\item{vertices}{Mesh vertex coordinates (V x 3)}

\item{faces}{Face indices (F x 3, 1-based)}

\item{data}{Vertex data (V) or (V x K)}
}
\value{
Interpolated values
}
\description{
Barycentric interpolation on surface mesh
}
\keyword{internal}
