% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deformation_field.R
\name{deformation_field}
\alias{deformation_field}
\title{Create a deformation field from a morphism on a grid}
\usage{
deformation_field(morphism, grid, with_jacobian = TRUE, log = FALSE)
}
\arguments{
\item{morphism}{A Morphism (affine or warp)}

\item{grid}{A Grid object (see grid_spec/grid_from_data)}

\item{with_jacobian}{Logical; attach Jacobian determinant volume}

\item{log}{If TRUE, store log-Jacobian}
}
\value{
An array with dim = c(grid@dims, 3) and class "DeformationField"
}
\description{
Create a deformation field from a morphism on a grid
}
