% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_ingest.R
\name{detect_fnirt_def_type}
\alias{detect_fnirt_def_type}
\title{Detect FNIRT deformation type (relative vs absolute)}
\usage{
detect_fnirt_def_type(warp_path, sample_n = 200, threshold_mm = 5)
}
\arguments{
\item{warp_path}{Path to FNIRT warp file}

\item{sample_n}{Number of voxels to sample for heuristic}

\item{threshold_mm}{Threshold for displacement magnitude heuristic}
}
\value{
"relative" or "absolute"
}
\description{
Heuristically determines whether an FNIRT warp stores relative
displacements or absolute coordinates.
}
\examples{
\dontrun{
def_type <- detect_fnirt_def_type("warp.nii.gz")
}
}
