% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_helpers.R
\name{detect_transform_type}
\alias{detect_transform_type}
\title{Read transform from file into a Morphism}
\usage{
detect_transform_type(path, source_affine = NULL, target_affine = NULL)
}
\arguments{
\item{path}{Path or Morphism}

\item{type}{Optional type hint: ants_h5, ants, fsl, fsl_coef, afni, x5, linear,
fsl_affine, afni_affine, itk_affine, lta_affine}

\item{source}{Optional source id}

\item{target}{Optional target id}

\item{apply_affine}{Logical; for ANTs H5 files, whether to include embedded affine}

\item{...}{Passed to morphism constructors}
}
\value{
A Morphism or MorphismPath object
}
\description{
Read transform from file into a Morphism
}
