% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_ingest.R
\name{fsl_fsl_to_vox}
\alias{fsl_fsl_to_vox}
\title{Build FSL-to-voxel scaling matrix}
\usage{
fsl_fsl_to_vox(affine, dim = NULL)
}
\arguments{
\item{affine}{4x4 voxel-to-world affine}

\item{dim}{Optional image dimensions (required to apply FSL handedness swap)}
}
\value{
4x4 inverse scaling matrix
}
\description{
Build FSL-to-voxel scaling matrix
}
\keyword{internal}
