% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_ingest.R
\name{fsl_vox_to_fsl}
\alias{fsl_vox_to_fsl}
\title{Build voxel-to-FSL scaling matrix}
\usage{
fsl_vox_to_fsl(affine)
}
\arguments{
\item{affine}{4x4 voxel-to-world affine}
}
\value{
4x4 scaling matrix
}
\description{
Build voxel-to-FSL scaling matrix
}
\keyword{internal}
