% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{grid_spec}
\alias{grid_spec}
\title{Create a grid specification}
\usage{
grid_spec(dims, affine, domain = NULL)
}
\arguments{
\item{dims}{Integer vector of dimensions}

\item{affine}{4x4 voxel-to-world matrix}

\item{domain}{Optional domain hash}
}
\value{
Grid object
}
\description{
Create a grid specification
}
