% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deformation_field.R
\name{jacobian_det_field}
\alias{jacobian_det_field}
\title{Extract Jacobian determinant volume from a morphism on a grid}
\usage{
jacobian_det_field(morphism, grid, log = FALSE)
}
\description{
Extract Jacobian determinant volume from a morphism on a grid
}
