% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sampled_points}
\alias{sampled_points}
\title{Construct sampled points}
\usage{
sampled_points(coords)
}
\arguments{
\item{coords}{Numeric matrix with 3 columns (N x 3)}
}
\value{
SampledPoints object
}
\description{
Construct sampled points
}
