% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphism.R
\name{SurfToSurfMorphism}
\alias{SurfToSurfMorphism}
\title{Create a surface-to-surface morphism}
\usage{
SurfToSurfMorphism(
  source,
  target,
  method = c("sphere", "area", "sulc"),
  source_sphere = "",
  target_sphere = "",
  cost = 1,
  method_tag = "anatomical"
)
}
\arguments{
\item{source}{Source surface domain hash}

\item{target}{Target surface domain hash}

\item{method}{Registration method: "sphere", "area", "sulc"}

\item{source_sphere}{Path to source sphere surface}

\item{target_sphere}{Path to target sphere surface}

\item{cost}{Path cost (default 1.0)}

\item{method_tag}{Method tag (default "anatomical")}
}
\value{
SurfToSurfMorphism object
}
\description{
Maps between surface meshes via spherical registration.
}
